Rename and move some code to perpare for refactor
