+++ internal/controller/supervisorconfig/generator/supervisor_secrets.go
@@ -6,14 +6,12 @@ package generator
 
 import (
 	"context"
-	"crypto/rand"
 	"fmt"
 
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	corev1informers "k8s.io/client-go/informers/core/v1"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/util/retry"
@@ -24,25 +22,9 @@ import (
 	"go.pinniped.dev/internal/plog"
 )
 
-// TODO: de-dup me when we abstract these controllers.
-const (
-	symmetricKeySecretType    = "secrets.pinniped.dev/symmetric"
-	symmetricKeySecretDataKey = "key"
-
-	symmetricKeySize = 32 // TODO: what should this be?
-)
-
 // generateKey is stubbed out for the purpose of testing. The default behavior is to generate a symmetric key.
 //nolint:gochecknoglobals
 var generateKey = generateSymmetricKey
-
-func generateSymmetricKey() ([]byte, error) {
-	b := make([]byte, symmetricKeySize)
-	if _, err := rand.Read(b); err != nil {
-		return nil, err
-	}
-	return b, nil
-}
 
 type supervisorSecretsController struct {
 	owner    *appsv1.Deployment
@@ -114,58 +96,6 @@ func (c *supervisorSecretsController) Sync(ctx controllerlib.Context) error {
 	c.setCache(newSecret.Data[symmetricKeySecretDataKey])
 
 	return nil
-}
-
-func isValid(secret *corev1.Secret) bool {
-	if secret.Type != symmetricKeySecretType {
-		return false
-	}
-
-	data, ok := secret.Data[symmetricKeySecretDataKey]
-	if !ok {
-		return false
-	}
-	if len(data) != symmetricKeySize {
-		return false
-	}
-
-	return true
-}
-
-func secretDataFunc() (map[string][]byte, error) {
-	symmetricKey, err := generateKey()
-	if err != nil {
-		return nil, err
-	}
-
-	return map[string][]byte{
-		symmetricKeySecretDataKey: symmetricKey,
-	}, nil
-}
-
-func generateSecret(namespace, name string, labels map[string]string, secretDataFunc func() (map[string][]byte, error), owner metav1.Object) (*corev1.Secret, error) {
-	secretData, err := secretDataFunc()
-	if err != nil {
-		return nil, err
-	}
-
-	deploymentGVK := schema.GroupVersionKind{
-		Group:   appsv1.SchemeGroupVersion.Group,
-		Version: appsv1.SchemeGroupVersion.Version,
-		Kind:    "Deployment",
-	}
-	return &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: namespace,
-			OwnerReferences: []metav1.OwnerReference{
-				*metav1.NewControllerRef(owner, deploymentGVK),
-			},
-			Labels: labels,
-		},
-		Type: symmetricKeySecretType,
-		Data: secretData,
-	}, nil
 }
 
 func (c *supervisorSecretsController) createSecret(ctx context.Context, newSecret *corev1.Secret) error {
